# static/swagger.json
{
    "swagger": "2.0",
    "basePath": "/api",
    "paths": {
...
        "/registration/trip/{trip_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Trip Identifier",
                    "name": "trip_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20Trip%20User"
                            }
                        }
                    }
                },
                "summary": "Get a list of all users registered for a specific trip",
                "operationId": "Get a list of all users registered for a specific trip",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
        "/registration/user/self": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Register the current user for a trip",
                "operationId": "Register the current user for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips the current user is registered for",
                "operationId": "Get a list of all trips the current user is registered for",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Cancel the current user's registration for a trip",
                "operationId": "Cancel the current user's registration for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Modify the current user's registration for a trip",
                "operationId": "Modify the current user's registration for a trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Registration%20without%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
        "/registration/user/{firebase_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID",
                    "name": "firebase_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Complete%20User%20Trip"
                            }
                        }
                    }
                },
                "summary": "Get a list of all trips a specific user is registered for",
                "operationId": "Get a list of all trips a specific user is registered for",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "registration"
                ]
            }
        },
...
        "/trip/{trip_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Trip Identifier",
                    "name": "trip_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20Trip"
                        }
                    }
                },
                "summary": "Get the specified trip",
                "operationId": "Get a specific trip",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Delete the specified trip",
                "operationId": "Delete the specified trip",
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Modify the specified trip",
                "operationId": "Modify the specified trip",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Trip%20without%20ID"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "trip"
                ]
            }
        },
...
        "/user/self": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20User"
                        }
                    }
                },
                "summary": "Get the current user's data",
                "operationId": "Get the current user's data",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "delete": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Delete the current user",
                "operationId": "Delete the current user",
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            },
            "put": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Update the current user's data",
                "operationId": "Update the current user's data",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User%20Self"
                        }
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
...
        "/user/{firebase_id}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "Firebase User ID",
                    "name": "firebase_id",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Complete%20User"
                        }
                    }
                },
                "summary": "Get the specified user",
                "operationId": "Get a specific user",
                "parameters": [
                    {
                        "name": "X-Fields",
                        "in": "header",
                        "type": "string",
                        "format": "mask",
                        "description": "An optional fields mask"
                    }
                ],
                "security": [
                    {
                        "Bearer Auth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        }
    },
    "info": {
        "title": "MountainMates API",
        "version": "0.1",
        "description": "This API facilitates hiking trips, participant registration, and trip organization."
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "Bearer Auth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
...
}